```text feat: Add react-select dependency for dropdown functionality
This commit adds the "react-select" dependency to the project's package.json file. The "react-select" library provides dropdown functionality, which will be used in the customer return item dashboard to implement a dropdown for selecting a bill ID. This enhancement improves the user experience by allowing sellers to easily search and select a specific bill ID when adding return items.