feat: Add "Products" link to DashSidebar
This commit adds a new link to the DashSidebar component, allowing users to navigate to the "Products" page in the dashboard. The link is represented by a Sidebar.Item component with an icon and the label "Products". This enhancement improves the navigation experience for users by providing easy access to the "Products" section of the dashboard.